Remove all usage of dynamic_cast<> (#39)
All the nodes already hold E57 type information, so we don't need to rely on RTTI. dynamic_cast<> is much slower (especially on MSVC) than custom simple type checks as it must be much more generic. Moreover, we already have those checks in place so usage of dynamic_cast<> is redundant.

+++ src/Common.h
@@ -77,9 +77,9 @@ namespace e57
 
    /// Create whitespace of given length, for indenting printouts in dump()
    /// functions
-   inline std::string space( int n )
+   inline std::string space( size_t n )
    {
-      return ( std::string( static_cast<size_t>( n ), ' ' ) );
+      return ( std::string( n, ' ' ) );
    }
 
    /// Convert number to decimal, hexadecimal, and binary strings  (Note hex
